fix: armv7 cross-arch build — usize overflow on 32-bit targets
CpuBackend::MAX_ALLOC: gate 4GB vs 512MB via #[cfg(target_pointer_width)].
Capabilities::memory_bandwidth: usize → u64 (50 GB/s overflows 32-bit).
Adapter fallback: 8GB → 2GB for memory_capacity (fits 32-bit usize).

Resolves primalSpring v0.9.17 genomeBin armv7 cross-arch gap.
Verified: cargo check --workspace --target armv7-unknown-linux-gnueabihf PASS.

Made-with: Cursor

+++ crates/runtime/universal/src/backends/wgpu_backend/initialization.rs
@@ -249,7 +249,7 @@ impl WgpuComputeUnit {
                 deterministic: false,
             },
             memory_capacity: memory_capacity as usize,
-            memory_bandwidth: bandwidth as usize,
+            memory_bandwidth: bandwidth,
             compute_throughput,
             optimal_batch_size: batch_size,
             supported_ops: vec![
